feature(user-reading-history): hook update
