+++ src/pages/index.ts
@@ -8,4 +8,4 @@ export * from './manga/manga.screen'
 export * from './NewChapter/NewChapter.screen'
 export * from './NewPost/NewPost.screen'
 export * from './Universe/Universe.screen'
-export * from './dashboard/Dashboard.screen'
+export * from './Dashboard/Dashboard.screen'
